fix(Rust): incorrect code block signature for rust
A code block will correctly syntax highlight Rust code with the `rust`
code block signature, not `rs`; this is understandably confusing because
the file extensions differ here.

Perhaps it would be worth renaming the `SupportedFileType` enum to
something like `SupportedCodeBlock` to more accurately reflect what it
is matching later on.