feat(Lib): Add support CRLF line endings closes #36
If a CRLF line ending is detected in the source text, it will be used in
the rest of the embedme tool for matching line endings and for joining
lines.

There is an assumption that all line endings are the same throughout the
file (i.e. not mixed CRLF and LF) and also that the line endings are the
same for the source file, and the code files which will be embedded in
the source file (i.e. CRLF + CRLF, or LF + LF).

Mixed endings may result in mixed endings throughout the file or other
undocumented behavior.